2.5.0
Optimize model inference and update version to 2.5.0

Enhancements in `AIManager.cs` include improved error handling and optimizations in the model inference process, with better checks for `_ioBindingInitialized` and `_reusableTensor`. The handling of the `frame` object has been streamlined, and the initialization of `inputArray` is now ensured.

In `MathUtil.cs`, the `BitmapToFloatArrayInPlace` method has been updated to accept a variable `imageSize`, improving flexibility. Additional checks for bitmap dimensions and refined parallel processing logic enhance performance for small images.

The `AboutMenuControl.xaml` file has been updated to reflect the new version "v2.5.0" in the UI.